refact: 커스텀 훅 useNotiModal로 분리
#64
