+++ src/components/rightSide/user/style.tsx
@@ -25,7 +25,7 @@ export const KebabIcon = styled(MdOutlineMoreVert)`
   cursor: pointer;
 `;
 
-export const EmptyInviteIcon = styled(VscBell)`
+export const EmptyNotiIcon = styled(VscBell)`
   width: 16px;
   padding: 2px;
   height: 100%;
@@ -33,7 +33,7 @@ export const EmptyInviteIcon = styled(VscBell)`
   cursor: pointer;
 `;
 
-export const NewInviteIcon = styled(VscBellDot)`
+export const NewNotiIcon = styled(VscBellDot)`
   width: 16px;
   padding: 2px;
   height: 100%;
