+++ src/modal/NotificationModal.tsx
@@ -1,7 +1,7 @@
 import * as S from "./layout/style";
 import React from "react";
 import { getSocket } from "socket/socket";
-import { NotiType } from "@rightSide/user/UserInfo";
+import { NotiType } from "hooks/useNotiModal";
 
 type modalProps = {
   close: () => void;
@@ -14,23 +14,28 @@ function NotificationModal(props: modalProps) {
     console.log(e.currentTarget.id);
     socket.emit("joinChatRoom", {
       roomId: Number(e.currentTarget.id),
-    })
+    });
     props.close();
-  }
+  };
 
   return (
     <S.NotificationLayout>
       <h3> 알림 </h3>
       <S.NotiContent>
-        {props.notiList.length > 0
-          ? props.notiList.map((noti) => {
+        {props.notiList.length > 0 ? (
+          props.notiList.map((noti) => {
             return (
               <>
-                <S.Span onClick={joinHandler} id={`${noti.chatId}`}>{noti.title}</S.Span><br />
+                <S.Span onClick={joinHandler} id={`${noti.chatId}`}>
+                  {noti.title}
+                </S.Span>
+                <br />
               </>
-            )
+            );
           })
-          : <span>새로운 알림이 없습니다.</span>}
+        ) : (
+          <span>새로운 알림이 없습니다.</span>
+        )}
       </S.NotiContent>
       <S.BtnWrapper>
         <S.ModalButton2 onClick={props.close}>확인</S.ModalButton2>
